add collection detail network request implementation

+++ app/src/main/java/com/example/adrianwong/shopifychallenge/repository/ICollectionDetailsRepository.kt
@@ -5,6 +5,6 @@ import com.example.adrianwong.shopifychallenge.datamodels.Product
 import com.example.adrianwong.shopifychallenge.datamodels.Result
 
 interface ICollectionDetailsRepository {
-    suspend fun getCollectionDetails(collectionId: Int, page: Int): Result<Exception, List<CollectionDetails>>
+    suspend fun getCollectionDetails(collectionId: Long, page: Int): Result<Exception, List<CollectionDetails>>
     suspend fun getProducts(ids: String, page: Int): Result<Exception, List<Product>>
 }
